+++ tests/helpers/index.ts
@@ -9,3 +9,4 @@ export * from "./inspector";
 export * from "./debug";
 export * from "./isolation";
 export * from "./assertions";
+export * from "./extractors";